Añadida pequeña verificacion en parametros de umbrales.

+++ image/image_filters.h
@@ -11,6 +11,6 @@ typedef struct {
 Image* image_to_gray_scale(Image* img);
 Pixel mask_image_pixel(Image* img, int i, int j, Mask* mask, int* success);
 Image* laplace_filter(Image* img, Mask* mask);
-Image* binarize_image(Image* img, uint8_t threshold);
+Image* binarize_image(Image* img, int threshold);
 void free_mask(Mask* mask);
 #endif //IMAGE_PIPELINE_IMAGE_FILTERS_H
